fix: move enterprise scim token helper
